week_2_workflow_orchestration/videos/make_docker_block.py
from prefect.infrastructure.docker import DockerContainer

docker_block = DockerContainer(
    image="froukje/prefect:zoom",
    image_pull_policy="ALWAYS",
    autoremove=True
    )

docker_block.save("zoom", overwrite=True")
